Don't overwrite secrets.env

+++ set-secrets.sh
@@ -3,8 +3,7 @@ set -euf -o pipefail
 # set -x
 
 
-echo "Falls du das wirklich machen willst, dann kommentier nachfolgende Zeile aus!"
-#exit 1
+test -f secrets.env && (echo "Please delete secrets.env and rerun." && exit 1)
 
 # stop and remove all containers, otherwise we can't pass the new parameters as environment variables
 docker-compose down
